+++ vida/resources/views/livewire/centros/selector-prestaciones-centro-modal.blade.php
@@ -1,0 +1,3 @@
+<div class="p-1">
+    @livewire('centros.selector-prestaciones-centro', ['centroId' => $centroId])
+</div>